requirements.txt
Pillow
argparse
base64
git+https://github.com/CompVis/taming-transformers.git
git+https://github.com/openai/CLIP.git
imageio
ipython
kornia
math
matplotlib
numpy
omegaconf
os
pathlib
scipy
subprocess
sys
torch
torchmetrics
torchvision
tqdm
urllib
zipfile
ftfy
clean-fid
diffusers